requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blis==1.0.1
catalogue==2.0.10
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cookiecutter==1.7.3
cymem==2.0.8
decorator==5.1.1
distro==1.9.0
Django==4.2
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.5.1
django-crum==0.7.9
django-extensions==3.2.3
django-livereload-server==0.3.2
django-tailwind==3.1.0
django-widget-tweaks==1.4.8
es_core_news_md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.8.0/es_core_news_md-3.8.0-py3-none-any.whl#sha256=478b8bb3f3e8eb149192f7d80e7cd64f990b7e9bccbc5329df41a57e86326be2
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl#sha256=e451a83d6df79b87e9eed0cb553f03e99e36a3bab18a7b79f0dcfd1fdf875e12
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gast==0.6.0
gensim==4.3.3
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.7
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jinja2-time==0.2.0
jiter==0.6.1
joblib==1.4.2
keras==3.5.0
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
livereload==2.7.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
multidict==6.1.0
murmurhash==1.0.10
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
openai==1.52.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
parso==0.8.4
pillow==10.3.0
poyo==0.5.0
preshed==3.0.9
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==4.25.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
RapidFuzz==3.10.1
rcssmin==1.1.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rjsmin==1.2.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
setuptools==74.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
tailwind==3.1.5b0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
text-unidecode==1.3
tf_keras==2.17.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.17.0
